fix
6. при перемиканні між сторінками зʼявляється зайвий відступ над хедером  - ???
Деякі сторінки скачуть не через хедер, в інших не помітив, потрібен приклад
8. відступ від аватарки юзера за макетом повинен складати 18рх, наразі менше

На макеті 14px, але ок переродив на 18

+Полагодив СВГ на мобілца
+Передивився ще раз всі модалки на всіх переломах
+Переробив навігацю на ul
