+++ src/components/header/UserProfile/UserProfile.module.css
@@ -3,7 +3,7 @@
   height: auto;
   display: flex;
   align-items: center;
-  gap: 14px;
+  gap: 18px;
 }
 
 .avatarHeaderImg {
@@ -33,8 +33,8 @@
   justify-content: center;
   box-shadow: rgba(0, 0, 0, 0.1);
   border-radius: 30px;
-  top: 45px;
-  left: 40%;
+  top: 60px;
+  right: 30px;
   z-index: 11111;
   background-color: rgba(250, 250, 250, 1);
 }
@@ -93,8 +93,8 @@
 @media screen and (min-width: 768px) {
   .UserLogoModal {
     margin-top: 30px;
-    top: 13%;
-    left: 70%;
+    top: 60px;
+    right: 120px;
     width: 177px;
     height: 134px;
     gap: 32px;
@@ -104,7 +104,7 @@
     width: 480px;
     height: 215px;
     top: 60px;
-    left: 40%;
+    right: 120px;
   }
   .UserLogoutModalText {
     margin-bottom: 32px;
@@ -130,7 +130,7 @@
     height: 215px;
     padding: 50px;
     top: 60px;
-    left: 55%;
+    right: 130px;
   }
 
   .UserLogoutModalText {
